Skeleton for new unit tests of SimulationUplink.
